Finished both parts.
I really liked my solution for today. The problem however is that
it is by far the slowest day so far. For both parts separately as
well as parsing it averages around one second. The fault is most
probably with my sparse representation of the problem. But if you
create a 2D array to keep track of all points you have trouble
with the ever expanding grid. Just pre allocating enough padding
from the start is not what I want.