Made some improvements.
Now it runs at around 400 ms. It's still not great, but seems to
approach the limit of how good it can be with the sparse representation.
Could maybe convert the counts hash map to a matrix as it is bounded
every iteration and very dense. Also implement my own hasher.