Finished both parts.
Forgot the problem so the first time I solved it with the wrong
game (disregaring board state). Then I used iteration to fill in
the array but now I used recursion. Iteration should be faster,
but trickier since you need to find the order. Since it already
runs in around 3-5ms I will not improve it any more.