Finished part 2.
Also remade part 1 to use the same result, but crop it to the smaller
region. The bounding could be made quicker by just removing how much
overlap things have with -50..50. As now I create a lot of cuboids
which won't be needed for anything else. But it was easy and ok fast.